Initial checkin.
This repository contains a copy of the actual remote execution related
tools from Bazel Buildbarn. They have been adjusted to be built on top
of bb-storage.

This repository doesn't include copies of the original deployments and
documentation. These will be moved into separate repos. We should still
add a README to this repository, though.