Add a utility function for generating bb-browser URLs
There will be a couple of new places in the codebase where we want to
create URLs that point to bb-browser. One example is the bb-scheduler
status page that can display all running actions.

This new function is also slightly more correct than the existing code.
It does, for example, work well with cases where bb-browser is hosted on
a subpath.