Remove support for default configuration values
These values are hardly ever what is desired. Furthermore, they are
often shared between applications, meaning that there is a reasonable
risk that users may specify them for one application, but not the other,
even though they need to remain in sync.

If there is a true desire for default values, I think we should work
towards providing a Jsonnet library that augments configuration files
with defaults.