Add a brand new scheduler that uses the new Remote Worker protocol
This new scheduler is aware of platform queues. It also includes a
status page that shows information about operations and workers known by
the scheduler. On these pages, it is also possible to kill operations
and drain workers.

There is no longer any need to provide a queue size. The scheduler is
willing to hold on to as many operations as are provided by clients. To
prevent an infinite amount of work from piling up, the scheduler
automatically purges operations that are abandoned by clients. This
means that clients need to keep on checking in for operations to remain
valid.