Fix output directory creation to be relative to working directory
REv2 mandates that output directories are created relative to
the working directory. Currently buildbarn creates these directories
relative to the build directory.

Construct output directory paths that are relative to the build
directory, as required by the filesystem package, but that are
relative to the working directory, as required by REv2