Split up Environment into BuildDirectory and Runner
When support for multiple runners for a single worker as added, we ran
into a pretty serious design flaw: the cardinality between the number of
build directories and runners had changed: 1:1 -> 1:*. Back then we sort
of ignored this issue, which led interesting bugs, such as cleaning of
the build directory not working properly in multi-runner setups. We
eventually worked around this by disabling this feature. Some TODOs in
cmd/bb_worker/main.go remained.

The goal of this change is to finally tackle that problem by basically
kicking out the Environment interface, which caused the 1:1 coupling. It
introduces separate BuildDirectory and Runner types. LocalBuildExecutor
now takes individual copies of each.

Because of this decoupling, we can also fix #20. Builds are no longer
performed in the build directory itself. Instead, they are run inside a
subdirectory called "root". This makes it possible to use the top level
directory for other purposes (e.g., storing stdout and stderr logs).

It also partially addresses #23. The runner protocol has been extended
to explicitly pass on the input root directory, so that a runner could
chroot into it.