Allow bb_runner to call a helper process to 'install' a temp directory
For certain kinds of workloads it is desirable to have a true /tmp
instead of using the TMPDIR environment variable. At the same time, it
is not desirable to use a single physical directory in which all of the
data shared between build actions is stored.

Operating systems like NetBSD provide a feature called 'magic symlinks'
that make it possible to create symbolic links that resolve to a
different location based on the credentials of the calling process
(e.g., the user ID). On those systems we could remove /tmp, replacing it
by such a symbolic link. This symbolic link could point to the temporary
directory created by bb_worker that is passed to bb_runner.

I'd rather not add that many OS dependent features to bb_{worker,runner}
to support these kinds of setups. Instead, let's add a gRPC protocol
that bb_runner can use to reach out to a helper process. We will likely
not ship any implementation; it is up to our users to implement them as
they see fit.