Create bb-runner-installer image
This image is for use as an init container on kubernetes, to allow
bb-runner to be used as a mixin to existing containers. Notably, this
allows users to pick any RBE image version, not just the one which
bb-remote-execution is currently selecting.

This turns out to be important bacause bazel-toolchains selects image
versions based on the version of bazel being run, and there is no
image which is compatible with all bazel versions.