Don't use the host environment in case the command environment is empty
Go's Cmd.Run() is a bit tricky to use, in that it distinguishes between
an empty and a nil slice when passing in environment variables. Because
currently don't explicitly populate Cmd.Env, but depend on append()
growing the list for us, we may pass in nil in case no explicit
environment variables are provided.

This commit solves this problem by explicitly calling make() to
construct a list of the right size. A nice side-effect is that we also
no longer need to perform reallocations.