Allow BuildDirectory operations to fail asynchronously
The naive build directory implementation that we have always
instantiates the input root up front. This means that we can't get any
errors later on. It's not unrealistic that people want to implement
versions of BuildDirectory that populate the input root in the
background (lazily, or simply asynchronously).

Let's extend BuildDirectory to take instances of the recently introduced
ErrorLogger interface. Extend LocalBuildExecutor to create a custom
context that gets terminated when the ErrorLogger gets invoked. This
ensures that build actions are terminated as soon as I/O errors occur.