Allow FilePool to be backed by a block device
FilePool is an allocator for temporary files. We don't really use it in
Buildbarn yet, but it's left as an extension point for people writing
custom implementations of BuildDirectory. So far we provide one
implementation of FilePool that's backed by memory, and one that's
backed by a directory.

I've observed in practice that the in-memory one can be too limited, and
that the directory backed one is too slow. I've managed to get
significant speedups by writing data to a block device directly.
Considering that we already have code to map block devices (for
LocalBlobAccess), it's pretty easy to get this working.