Upgrade to the latest version of bb-storage
This version of bb-storage removes the ContentAddressableStorage
interface. Let's go ahead and remove as many uses of this interface from
bb-remote-execution as well.

There are still a handful of places where we need this (for
GetDirectory(), GetFile() and PutFile()). Let's keep those working by
adding our own copy of ContentAddressableStorage that only declares
those functions. At some point we should consider tidying this up, for
example by decomposing this interface into a DirectoryCache and a
FilePlacer.