Don't create AC entries for failing build actions
At some point in the past, REv2 was extended to allow the AC to be used
for storing failures as well. We started doing this, but this is very
unhelpful in practice. It means that failures no longer have stable
bb-browser URLs.

Fall back to the old behaviour by creating uncached action result
entries in the CAS. Not just for infrastructure failures, but also for
actions that fail with a non-zero exit code.