Complete the removal of ContentAddressableStorage
ContentAddressableStorage is a bit of a wacky interface. It provides
three separate functions. Two out of three, GetFile() and GetDirectory(),
can be placed in separate interfaces: FileFetcher and DirectoryFetcher.

The third one, PutFile(), only has a single implementation. There is no
need to use an interface for that. It can just be a free function. Let's
move it into the only place where it's actually called:
NaiveBuildDirectory.

This change allows us to untangle some of the mess we have in
cmd/bb_worker/main.go. No longer do we need ReadWriteDecoupling-
ContentAddressableStorage. With this approach it also becomes more
evident how we're ever going to provide a single hardlinking cache, even
if multiple build directories are used. Actually implementing that is
left for another day.

Addresses: #17