Add trace span details
This inserts trace spans for the waitExecution stream, and annotates
the Execution spans with the details of what happened.

Annotating with response.Message in particularly powerful here because
that happens to contain the bb-browser URL corresponding to the
execution, which is sufficient to explore the details of the
execution. The rest of the attributes are primarily useful for
searching for the right trace.