Convert to OpenTelemetry
To better fit the otel model, a separate producer/consumer span is
produced for the interaction between bb_scheduler and bb_worker, and
the waitExecution spans are linked to it. This will result in tracing
implementations representing the deduplication accurately.

This won't build until the corresponding PR to bb-storage is
merged. There doesn't seem to be any neat way to do this transition -
I do it locally with a replace directive in go.mod.