Decouple BuildDirectory from filesystem.Directory
With the addition of BuildDirectory.UploadFile(), there is no longer any
need for the code built on top of BuildDirectory to open actual files on
disk. Let's copy just those methods from filesystem.Directory that we
need. That reduces the size of the BuildDirectory interface, thereby
making it easier to implement custom BuildDirectories.