Add LockPile, a simple deadlock avoidance algorithm
When locking multiple nodes in a graph, a deadlock avoidance algorithm
must be used to guarantee forward progress. This is a small helper type
that implements a simple algorithm: it orders locks by memory address.

For self-referential graphs, it may be annoying to deal with preventing
that the same lock is acquired twice. This type therefore keeps track of
a counter for every lock that is acquired, thereby making it safe to
lock the same graph node multiple times.

This helper type is going to used by the FUSE file system to be able to
provide fine-grained locking on a per-directory level.