Let bb_runner open the build directory lazily
Right now bb_runner opens the build directory once during startup. This
is impractical if the build directory is virtualized with a FUSE file
system. The file descriptor of the build directory may point to the
directory underneath the mount point, meaning that the runner process is
incapable of accessing the files exposed through FUSE.

This change adds a LazyDirectory that wraps around an existing instance
of filesystem.Directory and opens it for every individual access.