Add support for privilege separation to bb_runner
This change adds a new configuration option to bb_runner named
runCommandsAs. When set, build actions will be run as a different user.
This makes it safer to run untrusted actions on your cluster, as
processes can no longer tamper with the bb_runner process.

This option is still compatible with the process table cleaning setting.
It ensures that processes corresponding to the right user are
terminated.