Remove the writable path checker
It turns out that adding this functionality to bb_runner is a bad idea,
now that we also want to provide privilege separation. We want to
perform the writability checker as the same user that's going to run the
build actions. This is not going to help in case of privilege
separation, as the checking will be done as user root, as opposed to the
user running the build actions.

Let's kick this out, but in turn add something to make life easier for
people who want to do their own writability checker. Move the privilege
separation into a separate package, so that people can build their own
tool that spins up child processes in the same way that bb_runner does.