Add facilities for forwarding metadata from the scheduler to workers
We already have a method that allows runners to inject metadata into
ActionResults. This is what we use right now to log POSIX resource usage
information into bb_worker logs and show it on the bb_browser page.

This change makes it possible for bb_scheduler to do the same.