Add support for replacing temporary directories with symlinks
bb_worker already has a feature where it autommatically creates a
temporary directory for every build action that it runs. Especially for
FUSE-based workers this is useful, as that temporary directory is
subject to the same quotas as the build directory.

This new option adds the ability to let bb_runner remove /tmp, /var/tmp,
etc. at the start of every build action and replace it with symbolic
links pointing into the temporary directory created by bb_worker. This
makes it easier to get a handle on the amount of disk space consumed by
a build action.

Let's rename the temporary_directories to clean_temporary_directories to
make it explicit what this option does when compared to
symlink_temporary_directories.