No longer let bb_worker use the Command message provided by bb_scheduler
Now that REv2 has promoted platform properties from the Command message
to Action, we can simplify the scheduler: it no longer needs to load
Command messages from the CAS. This is great, as CAS messages may
sometimes be relatively large.

Let's start off with changing bb_worker to no longer rely on receiving a
Command message from the scheduler. It now needs to load this manually
at the start of every build action.