Add PrepopulatedDirectory.LookupAllChildren()
We want to add support to bb_clientd to be able to persist the contents
of bazel-out/ directories across restarts. In order to do this, we need
a method for traversing an entire directory hierarchy.

We already have FUSEReadDir(), but that function is FUSE specific. It
doesn't return any usable error messages, but also doesn't give us back
the types we need. Let's solve this by adding another method for this.