Let workers log the timeout of an action
Because the scheduler is going to be permitted to run actions with
custom timeout values, the value stored in the Action may no longer be
identical to what was provided by the client. Let's add some basic
logging to the workers so that we can derive what timeout value was
actually used.