bb_scheduler: Do prefix matching on instance names
One of the downsides of the existing scheduler is that it does exact
string matching on instance names. This means that it's not possible to
create pools of workers that are shared across projects that have
distinct instance names.

This change removes this limitation by switching to prefix matching of
instance names. To realize this, we add a new platform.Key type that can
be placed in a platform.Trie. The latter is not much more than a
per-platform map, containing an InstanceNameTrie.

Because we need to make sure that workers operate in the right context,
the Remote Worker protocol now sends the unmatched instance name suffix
to the worker at the start of every execution. This allows the workers
to reconstruct the full instance name and use that for all BlobAccess
traffic.