Scheduler: Add invocation ID based locality
This change extends InMemoryBuildQueue to keep track of the set of last
used invocation IDs associated with operations scheduled on workers.
Whenever new work comes in, clients will try to aggressively pick
idle workers that were used by the same invocation previously. This may
increase cache hits for certain kinds of workloads.

Furthermore, it adds a so-called worker invocation stickiness limit
option. This can be used to add a certain amount of 'unfairness' to the
scheduler. When operations end up getting queued, this option controls
to what extent operations from the previous invocation should be
preferred. Actual fairness in terms of worker count is still respected.
It's just that this option is used to determine which invocation needs
to be picked if the number of execution workers is tied.