SQS: Fix accidental reuse of the loop iterator
When doing some random browsing through the code, looking at places
where we create goroutines, I observed that this piece of code uses the
loop iterator inside the goroutine that gets created. This means that if
multiple messages are delivered, the goroutines may accidentally process
the wrong message.

As this code is used by the scheduler's lifecycle hook, this may under
heavy downscaling of the cluster cause us to accidentally kill workers
without draining them properly. Indeed, metrics show there is a ~1/10000
probability that this occurs.