Hook up the scheduler to the authorization API
Now that bb-storage provides a proper authorization API that's used for
BlobAccess, let's go ahead and extend the scheduler to use it as well.

For now, only use it to authorize GetCapabilities(), Execute() and
WaitExecution() (REv2) requests. In the future we can also use this for
the Remote Worker API and BuildQueueState.