Add nested fairness by turning invocations into a tree
bb_scheduler currently attempts to schedule operations fairly by
grouping them into invocations. The scheduler continuously tries to
keep the number of executing operations between invocations uniform.

For certain kinds of workloads, it is desirable to have nested forms of
fairness:

- Enforce fairness on a per user basis, but if a single user runs
  multiple builds, make each of those fair with respect to each other.

- Enforce fairness on the REv2 correlated invocations level, followed by
  the tool invocation ID.

- In case there is a large fixed cost w.r.t. setting up the execution
  environment: enforce fairness on the Docker container image level,
  followed by Bazel invocation ID.

To accomplish this, this commit changes InMemoryBuildQueue's invocations
from a flat heap/map to a tree. Instead of a single invocation key, each
operation may zero or more associated with it. The sequence of
invocations determines in which part of the invocation tree the
operation is placed.

Though this change complicates InMemoryBuildQueue slightly, there are
also some things that get simpler. For example, idle synchronizing
workers are now stored in exactly one queue instead of multiple. This is
now possible, because we can simply compute the invocation that's the
lowest common ancestor of all associated operations.

This change also patches up the bb_scheduler web UI to make the tree of
invocations traversable. Because the bookeeping of InMemoryBuildQueue is
now a bit more complete, we can display the list of executing workers
for a given invocation.