Let bb_worker upload outputs concurrently
Right now every worker thread uploads outputs sequentially. This is all
right when the number of outputs for an action is small and/or outputs
already happen to be present in the CAS. For workloads where actions
yield large Tree objects within many missing files, it does end up
becoming slow.

This change adds parallelism to uploads. To make sure that storage
doesn't get overloaded, there is a process-wide concurrency limit. This
means that if only one action is running on a worker, it can use up all
slots. If a worker is running many actions in parallel, they will all
get a portion of the slots.