Generalize our virtual file system code
bb_worker can be configured to perform all of its builds in a virtual
file system based on FUSE. Though this file system works pretty well on
Linux, on macOS it's more flaky. The main reason for this is that
macFUSE is simply not as robust and featureful as Linux's FUSE
implementation.

For this reason I'm currently experimenting with adding an NFSv4 server
to bb_worker, to act as an alternative to FUSE. This change acts as a
preparation for this. It takes the existing pkg/filesystem/fuse code,
moves it to pkg/filesystem/virtual, and refactors all of it to be
decoupled from go-fuse as much as possible. The bits that actually
depend on go-fuse are moved into a pkg/filesystem/virtual/fuse
subpackage.

This change also adds a new HandleAllocator API, acting as a replacement
for InodeNumberTree and EntryNotifier. The intention of this API is to
account for the difference between FUSE and NFSv4, where NFSv4 requires
that all nodes in the file system are resolvable by file handle.

A nice side effect of this change is that the generic virtual file
system code becomes buildable on any platform; not just ones supported
by go-fuse. This simplifies the build process.