NFSHandleAllocator: Return NFS4ERR_STALE where applicable
Right now we return NFS4ERR_BADHANDLE in case a request contains a
handle of a file that no longer exists. This is incorrect, as we should
only return this error when the handle is inherently incorrect (i.e.,
having a format that is never valid).

This change alters NFSHandleAllocator to return NFS4ERR_STALE in case
the provided handle has the right format, but does not refer to a file
that exists right now.