Allow VirtualClose() to close files multiple times
In the case of NFSv4 an open-owner only needs to call CLOSE once to
close a given file, even if it's been opened multiple times. I initially
tried to make it so that files would get closed whenever duplicate opens
by a single open-owner were detected, but this makes the code pretty
messy.

This change makes it so that the NFSv4 server can simply count the
number of times a given open-owner opened a file. When the CLOSE method
is called, we simply do a bulk close on the file.