Add st_mtim to InMemoryPrepopulateDirectory
It turns out that Bazel depends on having this information present for
bb_clientd. For regular files, Bazel doesn't really care about having
file timestamps, as it internally just uses the hash and size. For
directories it does depend on knowing timestamps. This can be seen by
looking at ActionMetadataHandler.fileArtifactValueFromStat(). For
directories, it calls FileArtifactValue.createForDirectoryWithMtime().

This change ensures that stat() calls done through FUSE and NFSv4 return
the value stored internally. In addition to that, it extends the Remote
Output Service protocol to provide a similar field. We will alter
bb_clientd to set this field accordingly.