Add bb_noop_worker
It is sometimes desirable to instruct Bazel to merely upload test into
the CAS without actually triggering its execution. The test can then be
inspected, or run interactively on any system of choice, potentially
with a debugger attached.

Bazel currently doesn't feature any upload command. Furthermore, even if
it did, it wouldn't know how to generate bb_browser URLs. Let's
therefore add a feature for this on our end.

This change adds a new worker type that does nothing more than return a
gRPC error for any action that is attempted to be scheduled on it. The
gRPC error message includes the bb_browser URL.