Reimplement LockPile to use the "Smart" algorithm (#1446)
When I implemented LockPile, I had to resort to acquiring locks in an
ordered fashion. The reason for this was that Go's sync.Mutex type only
provided a blocking Lock() operation; TryLock() didn't exist.

Go 1.18 changed this, as they finally included a TryLock() operation.
This means that we can reimplement LockPile to use what Howard Hinnant
calls the "Smart" algorithm, instead of the existing "Ordered"
algorithm. See the following page for details:

https://howardhinnant.github.io/dining_philosophers.html

The advantage of the "Smart" algorithm is that it should allow more
threads to make progress. With the "Ordered" algorithm you could in
theory have chains of threads all waiting for each other, which could,
for example, cause contention on the FUSE/NFSv4 root directory, even
though most blocking operations are performed in nested subdirectories.
With the "Smart" algorithm such chains should get broken up relatively
quickly.

A secondary benefit is that we can eliminate the use of reflection. We
currently need to use it to obtain the memory address of a lock, so that
we can apply a total ordering. The new implementation only cares about
equality. This also makes testing easier, as we no longer need to store
locks in a contiguous array. We can use regular mocks.