Call into macOS mount_nfs(8) to properly initialize NFSv4
There may be certain things that need to be initialized on the OS side
(kernel extensions, daemons, etc.) before we can safely call mount(2) to
create an NFSv4 mount.

Let's solve this by adding a no-op call to mount_nfs(8) to startup that
is known to fail quickly. That way the initialisation will take place as
expected.