This introduces a new bb_runner cleaner `run_command_cleaner`
The `run_command_cleaner` can be used to run an arbitrary command before
and after each action that is executed on an action runner. This is a
useful escape hatch to install custom behaviors that may not be
available by the buildbarn bb_runner natively.

Be warned, this does exec a new process, which has inherent performance
overhead in doing so.