Let bb_scheduler propagate public AuthenticationMetadata to bb_worker
Just like how bb_scheduler is capable of extracting the REv2
RequestMetadata message from the incoming request and forwarding it to
the worker through the auxiliary metadata field, this change extends it
to attach the public parts of the AuthenticationMetadata.

The effect of this change is that AuthenticationMetadata ends up getting
stored in the Action Cache and logged by the worker.