Add bb_virtual_tmp
bb_runner has the symlinkTemporaryDirectories option that can be used to
remove /tmp and replace it with a symbolic link pointing to the action's
temporary directory that bb_worker allocates in the build directory.
This has the advantage that /tmp is subject to the same quotas as the
build directory.

Unfortunately, this solution is insufficient for workers that support
concurrent execution, running on operating systems that don't provide
file system virtualisation facilities (such as macOS). On those systems
there is just a single /tmp that would need to be shared between build
actions.

This change adds a new daemon, bb_virtual_tmp, that provides a virtual
flie system that contains just a single symbolic link. The target of
this symbolic link can be altered through the
TemporaryDirectoryInstaller gRPC API, bb_runner's equally named
configuration option.

Because the virtual file system code has now been made aware of
credentials, it's possible to give each user its own symlink target. All
you need to do is ensure that both the gRPC server and the virtual file
system use the same authentication metadata format.

The idea is thus that the macOS installation is modified to let /tmp
point into the virtual file system offered by bb_virtual_tmp.
bb_virtual_tmp then forwards requests to the right build directory.