Let BlobAccessDirectoryFetcher use util.VisitProtoBytesFields()
On the client side we can also easily run into maximum message size
limit when accessing Tree objects through the virtual file system. This
change reimplements BlobAccessDirectoryFetcher in such a way that it
does streaming against Tree objects where possible.

A nice optimisation is that we can now slice Trees into individual
Directories without unmarshaling them. Only when the size of the
Directory matches that of the one that is requested, do we need to
duplicate and unmarshal. This should speed up ingestion of Tree objects
significantly.