Run operations with a higher expected execution duration first
It's often the case that the final stages of a build/test have poor
concurrency. The reason is that there are often a small number of
long-running actions that stretch the overall build time. Builds would
run faster if we made sure if those actions were spawned as soon as
possible.

Even though optimisations in this area are best made on the Bazel side,
there is already some low-effort improvement that we can make on the
Buildbarn side. Let's use the Initial Size Class Cache (ISCC) to compute
the expected execution duration of an operation, and use that to sort
queued operations in a given invocation.