NFSv4: Add client feature detection based on macOS version
Depending on the version of macOS used, it is possible to use NFSv4 over
UNIX domain sockets. Furthermore, it is sometimes possible to control
symlink target caching behaviour through a mount option, as opposed to
toggling a sysctl value.

This change adds logic to extract the build version of the running copy
of macOS, and based on the results use the correct mount options.