Add support for prefetching files to FUSE/NFSv4 based workers
FUSE/NFSv4 based workers currently fault in input files on demand,
without any readahead. This is because they depend on the build actions
to call into file system operations against the input root to access
data.

This change adds logic to bb_worker to record which files are being
accessed as part of execution. Upon completion, a Bloom filter is
calculated which is stored in the File System Access Cache (FSAC). The
next time a similar action executes, the Bloom filter is reloaded,
allowing the worker to launch a background job that 'races' with the
build action to fetch the data.

https://github.com/buildbarn/bb-adrs/blob/master/0010-file-system-access-cache.md