Ensure that running ARM64 binaries on x86-64 returns INVALID_ARGUMENT
If you try to run macOS ARM64 binaries on macOS x86-64 workers,
execution fails with gRPC status code INTERNAL. This is because macOS
returns EBADARCH instead of ENOEXEC.

As EBADARCH is specific to macOS, we can't list it in generic code. Add
a separate source file for macOS specific bits.