Make prefetching work with NFSv4
When implementing this code, there was already something fishy going on
that I couldn't pinpoint at the time:

- To detect directory access, we do not wrap the directories themselves.
  We wrap InitialContentsFetcher, which is used a layer deeper.
- To detect file access, we wrap the files themselves.

At first I thought this symmetry is all right, but this simply cannot
work for NFSv4. In the case of NFSv4 the HandleAllocator is responsible
for keeping a registry of live files in the file system. If we wrap
files, InMemoryDirectory will hold on to the wrapped instance, but the
HandleAllocator would not. This means that VirtualRead() calls issued
against files obtain from the registry would bypass the monitor. So the
layering of NativeLeaf is, from outermost to innermost:

- accessMonitoringNativeLeaf
- nfsStatelessNativeLeaf
- blobAccessCASFile

While we should have something that resembles this:

- nfsStatelessNativeLeaf
- accessMonitoringNativeLeaf
- blobAccessCASFile

This change solves this by making statelessHandleAllocatingCASFileFactory
responsible for wrapping blobAccessCASFile to have a read monitor
attached to it. This ensures that if VirtualRead() is called against
nfsStatelessNativeLeaf, it will always be subject to the monitor.