Let bb_worker use graceful termination
Now that bb-storage provides decent tools for running programs using
structured concurrency, we can patch up bb_worker to make full use of
it. Namely, we can add graceful handling of shutdown signals, so that we
can run currently executing actions to completion, and gracefully shut
ourselves down on the scheduler side.

To achieve the latter, we need to make some minor changes to Buildbarn's
Remote Worker protocol. Right now we can only set 'prefer_being_idle'
when we're executing. If we want to shut down gracefully, we also need
to be able to set this flag while idle. If we get interrupted while
doing a blocking Synchronize() call, we need to perform a retry against
the scheduler to ensure we didn't miss any task assignment. For that we
need to provide 'prefer_being_idle'.

With this change applied, it is a lot safer to do rolling upgrades of
workers. When running your cluster on Kubernetes, I would strongly
recommend making the following configuration changes:

- On your worker pods, set Kubernetes pod spec option
  terminationGracePeriodSeconds to a value that exceeds the
  maximumExecutionTimeout configured on your scheduler. This is
  necessary to ensure Kubernetes doesn't terminate your workers while
  it's still running actions to completion.

- Enable gRPC server option stopGracefully on your bb_runner processes.
  This ensures that Kubernetes does not shut down your bb_runner
  processes while bb_worker is trying to use it to run an action.