Generalize BuildQueueState.KillOperation()
For people writing custom autoscalers, it may be useful to be able to
send a message back to the scheduler in case the platform is invalid or
unusable. For example, a container_image platform property may have the
wrong format, or systems that are being requested don't exist. For that
it is desirable if there was a way where all queued operations belonging
to a platform without workers can be terminated at once.

Instead of adding a separate RPC for this, this change reworks the
existing KillOperation() RPC to support bulk termination of operations.
It now takes a filter, similar to some of the List*() operations.