bb_runner: Add support for injecting Xcode environment variables
When performing remote builds against systems that run macOS, it may be
desirable to call into copies of Xcode that are installed on the worker.
Xcode can be installed at an arbitrary location on the file system.
Furthermore, it is possible to install multiple copies, having different
versions.

To prevent the need for specifying absolute installation paths on the
client side, Bazel assumes that the remote execution service is capable
of processing the following environment variables:

- APPLE_SDK_PLATFORM
- APPLE_SDK_VERSION_OVERRIDE
- XCODE_VERSION_OVERRIDE

Based on those environment variables, the worker should make sure that
the following ones are set, which Xcode utilities like xcrun and
xcodebuild respect:

- DEVELOPER_DIR
- SDKROOT

This change adds a configuration option for bb_runner, allowing you to
specify a mapping between values of XCODE_VERSION_OVERRIDE and the
intended DEVELOPER_DIR. The value of SDKROOT is automatically set, based
on what is returned by "xcrun --sdk ${sdk_name} --show-sdk-path".

Both Bazel and other remote execution systems have similar logic for
performing these translations:

- https://github.com/bazelbuild/bazel/blob/master/src/main/java/com/google/devtools/build/lib/exec/local/XcodeLocalEnvProvider.java
- https://github.com/bazelbuild/bazel-buildfarm/blob/main/macos-wrapper.sh