Stop using deprecated REv2.0 fields as part of tests
REv2.0 uses the outputs_files and outputs_directories to specify which
files/directories should be captured after execution of an action
completes. REv2.1 uses a single output_paths field. This is better,
because it allows clients to schedule actions, not knowing what type of
file gets generated.

It is important that both schemes are supported, but that's already
abstracted away entirely by OutputHierarchy. Let's use the REv2.1
semantics as part of all other tests.