Tidy up VirtualClose()
If a single NFSv4 owner-owner attempts to open the same file multiple
times (e.g., due to hardlinking or because it opens the same file both
for reading and writing), we still only see a single CLOSE operation.

We solved this by letting the CLOSE operation call VirtualClose() with
the count of the number of times the file was opened. This is somewhat
convoluted, as we can also just close the redundantly opened file when
the transaction completed. This means we always call VirtualClose() with
a count of 1, allowing us to remove the parameter.