VFS: Allow files to track whether they are opened for reading/writing
Right now implementations of Leaf know whether they are being opened for
reading/writing, but can't accurately tell whether they still are at a
given point in time. This can be impractical at times, for example if a
lack of writers is used to determine whether a file needs to be flushed
to underlying storage.

This change solves that by letting VirtuaClose() also take a ShareMask.