Allow the Remote Output Service to not report a file's digest
When bb_clientd's FUSE file system is configured to use write-back
caching, then there's no guarantee that if BatchStat() is called,
bb_clientd has access to the contents of the latest version of a given
file. This would cause it to report an incorrect digest.

This change extends the Remote Output Service protocol to document that
implementations such as bb_clientd may omit the digest in case they
detect scenarios in which an incorrect digest would be reported, such as
one or more writable file descriptors existing.

Implementing this actual logic remains to be done.