Allow PageRank to be used independently of the ISCC
For certain workloads it may not be necessary to use PageRank to
automatically pick an optimally sized worker. Still, it may be useful to
enable the Initial Size Class Cache (ISCC). That way it's possible to
observe timing statistics in bb_browser.

This change adds a simple StrategyCalculator that is functionally
equivalent to FallbackAnalyzer. It always runs actions on the smallest
size class, falling back to running it on the largest.