PoolBackedFile: Suppress digests when file is opened for writing
When using Remote Output Service, there may be cases where Bazel asks
bb_clientd to compute a file's digest, but bb_clientd may not always be
able to give a decent answer. For example, when a locally created file
is opened for writing, the kernel may still hold on to unflushed pages
of data.

This change extends PoolBackedFile to simply not return a digest in
those cases. It is then up to Bazel to compute the digest manually. As
Bazel reads file contents through the page cache, it will always observe
the latest version of the file.