Let bb_runner check that it can access the build directory
I've seen multiple reports of people on Slack trying to switch to FUSE,
but running into issues with mount propagation. This currently causes
the worker and runner to come online and pick up work, but the execution
of those actions will fail due to the runner being unable to open
stdout/stderr.

Let's be a bit more aggressive here and change this to cause readiness
checks to fail. This causes the worker to stop synchronizing against the
scheduler. This is great, because it means that if FUSE/NFSv4 were to
become unhealthy for some odd reason, it wouldn't cause the worker to
repeatedly send failed execution responses back to the scheduler.