Add support for uploading output directories as Directory messages
https://github.com/bazelbuild/remote-apis/pull/258 added support to REv2
for storing output directories in the form of individual Directory
messages, as opposed to using Trees. This change adds partial support
for it. Namely, we always upload Trees, but additionally upload separate
Directory messages if requested by the client.

Furthermore, we add a configuration option to force enable it. Some
users may desire this for the reasons documented in bb_worker.proto.