Log action errors for build directory cleanup
There was no error reporting for failure to clean up the build
directories and this would fail creation of the same action the next
time it is executed. This is an unexpected error and we have not needed
it before, but on Windows some files may be locked. We cannot return an
error to the client, as it may retry the action and then fail directory
creation, and keep retrying.