Command Creator decorator to remap executables
This is useful for Bazel that does not handle the shell executable well
for multi-platform builds. The default values are applied for each
platform, but if a runner uses a non-default location of bash. Bazel has
no way to change that for just a single platform. A workaround is to
remap the executable path in the runner's CommandCreator.