Share path resolution code for Windows
The path lookup is symmetric on all platforms and more code can be
shared. The only difference is that the command and argv0 must use
backslashes on Windows. The windows code can reuse the same path lookup
code, just with the modification that the paths sent from a Linux client
use forward slashes.