Add support for delaying uploads until output files are closed
The kernel may sometimes close files asynchronously. This means that if
an action completes and uploading of output files starts, the files may
still be opened for writing. This may cause the kernel to still have
dirty pages, which don't end up getting uploaded.

This change solves that by adding some logic to our mutable files
implementation to let UploadFile() block until files are no longer
opened for writing. As there are certain circumstances under which files
could be held open indefinitely, we put a limit in place that needs to
be specified in the configuration file.

This issue was observed by using the recently added server logs feature.
I'm using this to attach core dumps to execution results. These would
sometimes be zero bytes in size, due to them being written
asynchronously.