Build multi-architecture container images
This makes it possible to use most of our images on ARM64 based systems.
We still need to spend some time to actually make bb_runner_installer a
multi-architecture image. Maybe the easiest thing is to just rewrite it
in Go, and use //go:embed to embed copies of binaries in there that are
needed for execution.